fix: harden quadlet export and clear 0.13.0 audit findings (#197)
Closes #196.

## Security (HIGH)
Quadlet export (`podup generate quadlet --output`) treated
parsed-compose keys and field values as trusted:
- service/network/volume keys were used verbatim as unit file names →
path traversal / arbitrary write;
- field values were written without stripping control characters →
newline-based systemd directive injection.

Fixed by reducing unit file-name stems to a safe single path component,
stripping ASCII control characters from every dynamic value, and a
fail-closed guard in the writer. +5 tests.

## Correctness / release
- Lower MSRV 1.86 → 1.85 (the idna/icu chain that forced 1.86 was
removed); verified `cargo +1.85.0 test --lib` passes. Updated
`Cargo.toml`, CI `msrv`, `debian/control`, the debian-build comment and
the packaging doc. Debian trixie can now build the package.
- Add the 0.13.0 `debian/changelog` stanza.

## Structure
- Split the four files over the 500-line hard limit (quadlet, main.rs,
engine container_config, libpod spec) into submodules; rename
`container_misc` → `container_fields`. No file now exceeds 500 lines.

## Tests
- Make the watch integration tests deterministic: replace fixed sleeps
with polling on the observable effect (new test-only
`test_exec_capture`), with real assertions.
- Add unit tests for `extends/merge.rs`.

## Docs
- Add `debian/podup.1` (installed via `dh_installman`) and
`docs/commands.md`; fix the README's stale library tag and orphaned doc
link; correct the packaging doc's MSRV.

## Verification
`cargo fmt --check` clean, `cargo clippy --all-targets --all-features`
clean, 482 lib + 241 integration tests pass, builds on 1.85.

---------
