+++ Cargo.toml
@@ -2,7 +2,7 @@
 name = "podup"
 version = "0.13.0"
 edition = "2021"
-rust-version = "1.86"
+rust-version = "1.85"
 description = "Translate and run docker-compose files on rootless Podman"
 license = "Apache-2.0"
 repository = "https://github.com/Glyndor/podup"
