+++ debian/control
@@ -2,7 +2,7 @@ Source: podup
 Section: utils
 Priority: optional
 Maintainer: Glyndor <75870284+Jaro-c@users.noreply.github.com>
-Build-Depends: debhelper-compat (= 13), cargo, rustc (>= 1.86)
+Build-Depends: debhelper-compat (= 13), cargo, rustc (>= 1.85)
 Standards-Version: 4.7.0
 Homepage: https://github.com/Glyndor/podup
 Vcs-Git: https://github.com/Glyndor/podup.git
