+++ debian/podup.manpages
@@ -1,0 +1,1 @@
+debian/podup.1
